add trojan

+++ start.sh
@@ -1,8 +1,10 @@
 #!/bin/bash -ex
 : ${DOMAIN:?is not given.}
 : ${UUID:?is not given.}
+: ${TROJAN_PASSWORD:?is not given.}
 [[ -z $(nslookup $DOMAIN | grep $(curl -s http://checkip.amazonaws.com)) ]] && echo "fail to validate domain" && exit 1
 envsubst '${UUID}' < config.json.template > ./appdata/etc/v2ray/config.json
+envsubst '${TROJAN_PASSWORD}' < trojan.config.json.template > ./appdata/config/config.json
 envsubst '${DOMAIN}' < docker-compose.yml.template > docker-compose.yml
 envsubst '${DOMAIN}' < v2ray.subdomain.conf.template > appdata/config/nginx/proxy-confs/v2ray.subdomain.conf
 docker-compose down
